requirements.txt
flask
slack_sdk
supabase
python-dotenv
gunicorn